Tweak the request timeouts
After a while the wundasmart box freezes up. My suspiction is that
connections are not being closed correctly when a timeout occurs.
Instead of timing out based on the total request time and attempting to
close the connection, use the lower level socket timeouts to see if that
works better.