Fix getting current operation mode for water heater (#28)
State and operation_mode are the same thing for the WaterHeaterEntity,
and the current operation mode (which self.state returns) has to be an
operation mode that can also be selected.

When a manual override has been set try to determine the most sensible
operation mode to display. Otherwise, use on/off and add those to the
list of operation modes allowed. Since we can't turn the water heater on
or off without a duration, alias on/off to the longest preconfigured
boost/off modes.

Fixes #27