Use long lived TCP connection to hub and add heartbeat/ping sensor (#50)
* Use a long lived TCP connection to the hub

Instead of creating a new connection for each HTTP request, keep
a long lived TCP connection alive and use it for all requests.

This should use fewer resources on the hub and hopefully avoid
stability problems that can occur when the hub has been up
for many months.

* Add heartbeat sensor that pings the wunda hub

Periodically ping the hub to check it's still responding.

This should help keep the hub's network interface awake.

* Add missing requirement for tests

* Remove unnecessary logging

* Treat empty values from syncvalues as None

Elsewhere it's assumed that unavailable values are either not set
in state dictionary or are None. Where empty strings are used it's
logging some warnings about float conversions which aren't needed.